+++ satella/coding/decorators/decorators.py
@@ -17,7 +17,7 @@ def _TRUE(x):
 Queue = tp.TypeVar('Queue')
 
 
-def queue_get(queue_getter: tp.Callable[[object], Queue], timeout: tp.Optional[float] = None,
+def queue_get(queue_getter: tp.Union[str, tp.Callable[[object], Queue]], timeout: tp.Optional[float] = None,
               exception_empty=queue.Empty,
               queue_get_method: tp.Callable[[Queue, tp.Optional[float]], tp.Any] =
               lambda x, timeout: x.get(