+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.11.6
+
+* added BogusTerminableThread
+* added `method_to_execute_on_empty` to `queue_get`