+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.7
+
+* fixed typing in queue_get