+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.11.7
 
 * fixed typing in queue_get
+* fixed a bug wherein single-op predicate would discard operations made on it so far