+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.8
+
+* added is_empty