+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.11.8
 
 * added is_empty
+* added `x.has_p`