+++ CHANGELOG.md
@@ -1,4 +1,2 @@
-# v2.11.9
-
-* changed `has_p` to `has`, deprecated `has_p`
+# v2.11.10